% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{j_patch_apply}
\alias{j_patch_apply}
\alias{j_patch_from}
\title{Patch or compute the difference between two JSON documents}
\usage{
j_patch_apply(data, patch, as = "string", ...)

j_patch_from(data_x, data_y, as = "string", ...)
}
\arguments{
\item{data}{JSON character vector, file, URL, or an \emph{R} object to
be converted to JSON using \code{jsonline::fromJSON(data, ...)}.}

\item{patch}{JSON 'patch' as character vector, file, URL, or \emph{R}
object.}

\item{as}{character(1) return type; \code{"string"} returns a JSON
string, \code{"R"} returns an \emph{R} object using the rules in
\code{as_r()}.}

\item{data_x}{As for \code{data}.}

\item{data_y}{As for \code{data}.}
}
\value{
\code{j_patch_apply()} returns a JSON string or \emph{R} object
representing 'data' patched according to 'patch'.

\code{j_patch_from()} returns a JSON string or \emph{R} object
representing the difference between 'data_x' and 'data_y'.
}
\description{
\code{j_patch_apply()} uses JSON Patch
\url{https://jsonpatch.com} to transform JSON 'data' according the
rules in JSON 'patch'.

\code{j_patch_from()} computes a JSON patch describing the
difference between to JSON documents.
}
\examples{
data_file <- system.file(package = "rjsoncons", "extdata", "patch_data.json")

## add a biscuit
patch <-
    '{"op": "add", "path": "/biscuits/1", "value": {"name": "Ginger Nut" }}'
j_patch_apply(data_file, patch, as = "R") |> str()

## add a biscuit and choose a favorite
patch <- '[
    {"op": "add", "path": "/biscuits/1", "value": {"name": "Ginger Nut" }},
    {"op": "copy", "from": "/biscuits/0", "path": "/best_biscuit"}
]'
biscuits <- j_patch_apply(data_file, patch)
as_r(biscuits) |> str()

j_patch_from(data_file, biscuits, as = "R") |> str()

}
