% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Version of jsoncons C++ library}
\usage{
version()
}
\value{
\code{version()} returns a character(1) major.minor.patch
version string .
}
\description{
\code{version()} reports the version of the C++ jsoncons
library in use.
}
\examples{
version()

}
