% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsoncons.R
\name{jsonpath}
\alias{jsonpath}
\alias{jmespath}
\alias{jsonpointer}
\alias{jsonpivot}
\title{Query the jsoncons C++ library}
\usage{
jsonpath(data, path, object_names = "asis", as = "string", ...)

jmespath(data, path, object_names = "asis", as = "string", ...)

jsonpointer(data, path, object_names = "asis", as = "string", ...)

jsonpivot(data, object_names = "asis", as = "string", ...)
}
\arguments{
\item{data}{an \emph{R} object. If \code{data} is a scalar (length 1)
character vector, it is treated as a single JSON
string. Otherwise, it is parsed to a JSON string using
\code{jsonlite::toJSON()}. Use \code{I()} to treat a scalar character
vector as an \emph{R} object rather than JSON string, e.g., \code{I("A")}
will be parsed to \verb{["A"]} before processing.}

\item{path}{character(1) jsonpath or jmespath query string.}

\item{object_names}{character(1) order \code{data} object elements
\code{"asis"} (default) or \code{"sort"} before filtering on \code{path}.}

\item{as}{character(1) return type. \code{"string"} returns a single
JSON string; \code{"R"} returns an \emph{R} object following the rules
outlined below.}

\item{...}{arguments passed to \code{jsonlite::toJSON} when \code{data} is
not a scalar character vector. For example, use \code{auto_unbox = TRUE} to automatically 'unbox' vectors of length 1 to JSON
scalar values.}
}
\value{
\code{jsonpath()}, \code{jmespath()} and \code{jsonpivot()} return a
character(1) JSON string (\code{as = "string"}, default) or \emph{R}
object (\code{as = "R"}) representing the result of the query.
}
\description{
\code{jsonpath()} executes a query against a JSON string
using the 'jsonpath' specification

\code{jmespath()} executes a query against a JSON string
using the 'jmespath' specification.

\code{jsonpointer()} extracts an element from a JSON string
using the 'JSON pointer' specification.

\code{jsonpivot()} transforms a JSON array-of-objects to
an object-of-arrays; this can be useful when forming a
column-based tibble from row-oriented JSON.
}
\details{
\code{jsonpivot()} transforms an 'array-of-objects' (typical when the
JSON is a row-oriented representation of a table) to an
'object-of-arrays'. A simple example transforms an array of two
objects each with three fields \code{'[{"a": 1, "b": 2, "c": 3}, {"a": 4, "b": 5, "c": 6}]'} to an object with with three fields, each a
vector of length 2 \code{'{"a": [1, 4], "b": [2, 5], "c": [3, 6]}'}. The
object-of-arrays representation corresponds closely to an \emph{R}
data.frame or tibble, as illustrated in the examples.
}
\examples{
json <- '{
  "locations": [
    {"name": "Seattle", "state": "WA"},
    {"name": "New York", "state": "NY"},
    {"name": "Bellevue", "state": "WA"},
    {"name": "Olympia", "state": "WA"}
  ]
 }'

## return a JSON string
jsonpath(json, "$..name") |>
    cat("\n")

## return an R object
jsonpath(json, "$..name", as = "R")

## create a list with state and name as scalar vectors
lst <- jsonlite::fromJSON(json, simplifyVector = FALSE)

## objects other than scalar character vectors are automatically
## coerced to JSON; use `auto_unbox = TRUE` to represent R scalar
## vectors in the object as JSON scalar vectors
jsonpath(lst, "$..name", auto_unbox = TRUE) |>
    cat("\n")

## a scalar character vector like "Seattle" is not valid JSON...
try(jsonpath("Seattle", "$"))
## ...but a double-quoted string is
jsonpath('"Seattle"', "$")

## use I("Seattle") to coerce to a JSON object ["Seattle"]
jsonpath(I("Seattle"), "$[0]")      |> cat("\n")

## different ordering of object names -- 'asis' (default) or 'sort'
json_obj <- '{"b": "1", "a": "2"}'
jsonpath(json_obj, "$")             |> cat("\n")
jsonpath(json_obj, "$.*")           |> cat("\n")
jsonpath(json_obj, "$", "sort")   |> cat("\n")
jsonpath(json_obj, "$.*", "sort") |> cat("\n")

path <- "locations[?state == 'WA'].name | sort(@)"
jmespath(json, path) |>
    cat("\n")

## original filter always fails, e.g., '["WA"] != 'WA'
jmespath(lst, path)  # empty result set, '[]'

## filter with unboxed state, and return unboxed name
jmespath(lst, "locations[?state[0] == 'WA'].name[0] | sort(@)") |>
    cat("\n")

## automatically unbox scalar values when creating the JSON string
jmespath(lst, path, auto_unbox = TRUE) |>
    cat("\n")

## jsonpointer 0-based arrays
jsonpointer(json, "/locations/0/name")

## document root "", sort selected element keys
jsonpointer('{"b": 0, "a": 1}', "", "sort", as = "R") |>
    str()

## 'Key not found' -- path '/' is searches for a 0-length key
try(jsonpointer('{"b": 0, "a": 1}', "/"))

json |>
    ## 'locations' is a array of objects with 'name' and 'state' scalars...
    jmespath("locations") |>
    ## ...pivot to a single object with 'name' and 'state' vectors...
    jsonpivot(as = "R") |>
    ## ... easily coerced to a data.frame or dplyr::tibble
    as.data.frame()

}
