% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsoncons.R
\name{version}
\alias{version}
\alias{jsonpath}
\alias{jmespath}
\title{Query the jsoncons C++ library}
\usage{
version()

jsonpath(data, path, auto_unbox = FALSE, ...)

jmespath(data, path, auto_unbox = FALSE, ...)
}
\arguments{
\item{data}{character(1) or list() Either a single JSON string or an R list
structure as obtained from \code{jsonlite::fromJSON}}

\item{path}{character(1) jsonpath or jmespath query string.}

\item{auto_unbox}{automatically \code{\link[jsonlite:unbox]{unbox()}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link[jsonlite:unbox]{unbox()}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{\link[=I]{I()}}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}

\item{...}{arguments passed on to \code{jsonlite::toJSON}}
}
\value{
\code{version()} returns a character(1) major.minor.patch version
string .

\code{jsonpath()} aand \code{jmespath()} return a character(1) json
string representing the result of the query.
}
\description{
\code{version()} reports the version of the C++ jsoncons library in
use.

\code{jsonpath()} executes a query against a json string using the
'jsonpath' specification

\code{jmespath()} executes a query against a json string sing the
'jmespath' specification.
}
\examples{
version()

json <- '{
  "locations": [
    {"name": "Seattle", "state": "WA"},
    {"name": "New York", "state": "NY"},
    {"name": "Bellevue", "state": "WA"},
    {"name": "Olympia", "state": "WA"}
  ]
 }'

jsonpath(json, "$..name") |>
    cat("\n")

jmespath(json, "locations[?state == 'WA'].name | sort(@)") |>
    cat("\n")

}
