% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsoncons.R
\name{version}
\alias{version}
\alias{jsonpath}
\alias{jmespath}
\title{Query the jsoncons C++ library}
\usage{
version()

jsonpath(data, path)

jmespath(data, path)
}
\arguments{
\item{data}{character(1) JSON string.}

\item{path}{character(1) jsonpath or jmespath query string.}
}
\value{
\preformatted{`version()` returns a character(1) major.minor.patch version
string .

`jsonpath()` aand `jmespath()` return a character(1) json
string representing the result of the query.
}
}
\description{
\preformatted{`version()` reports the version of the C++ jsoncons library in
use.

`jsonpath()` executes a query against a json string using the
'jsonpath' specification

`jmespath()` executes a query against a json string sing the
'jmespath' specification.
}
}
\examples{
version()

json <- '{
  "locations": [
    {"name": "Seattle", "state": "WA"},
    {"name": "New York", "state": "NY"},
    {"name": "Bellevue", "state": "WA"},
    {"name": "Olympia", "state": "WA"}
  ]
 }'

jsonpath(json, "$..name") |>
    cat("\n")

jmespath(json, "locations[?state == 'WA'].name | sort(@)") |>
    cat("\n")

}
